Ui improvements and add favicons
